Kernel & LibHardware: Move SMBIOS definitions to LibHardware
